% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Create Decision Tree Function}
\usage{
create_tree(edgelist)
}
\arguments{
\item{edgelist}{a list of edges with the following fields, in any order:
"from":             source node for this edge
"to":               target node for this edge
"name":             edge name (optional)
"outcome":          edge outcome, i.e. target node label (optional)
"type":             source target type (optional, case insensitive), options:
                    "D","decision" - decision node (square)
                    "C","chance" - chance node (circle)
                    "E","end" - end node (triangle, optional)
"probability":      edge probability
<payoff1>,<p2>,...: any number of cost/effect fields (optional)}
}
\value{
graph object with the appropriate properties and node types
}
\description{
This function allows you to create a decision tree object.
}
\examples{
tree <- create_tree(read.csv("branches.csv"))

}
