% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tree.R
\name{estimate_tree}
\alias{estimate_tree}
\title{Estimate Decision Tree Function}
\usage{
estimate_tree(tree, payoffs = NULL)
}
\arguments{
\item{tree}{a decision tree/graph object generated by create_tree()}

\item{payoffs}{a list of cost/effect column names to use during estimation}
}
\value{
list with tree outcomes (pathprobabs, payoffs, weighted payoffs)
}
\description{
This function estimates final payoffs from a decision tree object.
}
\examples{
results <- estimate_tree(tree, c("cost","effect"))

}
