% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot Decision Tree Function}
\usage{
plot_tree(tree, edge.label.display = "both",
  vertex.label.display = "end", edge.label.position = 0.33,
  vertex.size = 7, font.size = 3, plot.padding = "auto",
  color.root = "yellow", color.internal = "green", color.end = "red",
  theme = FALSE)
}
\arguments{
\item{tree}{a decision tree/tree object generated by create_tree()}

\item{edge.label.display}{edge label mode, options:
- "all","both": concatenate edge names and probabilities (default)
- "label": show edge names only
- "probability": show edge probabilities only
- "none": don't show edge labels}

\item{vertex.label.display}{vertex label mode, options:
- "all": show labels for all vertices
- "end": show labels only for end vertices (default)
- "none": don't show vertex labels}

\item{edge.label.position}{how far along the edge's straight part to display the label
([0-1], default = 0.33)}

\item{vertex.size}{size of vertices in mm (default = 7)}

\item{font.size}{font size in edge and vertex labels (default = 3)}

\item{plot.padding}{left and right padding to add to the plot, options:
- "auto": automatically set based on font size and max label length (default)
- c(left, right): manually set based on provided vector}

\item{color.root}{color of the root node (default = "yellow")}

\item{color.internal}{color of internal, i.e. not root or end nodes (default = "green")}

\item{color.end}{color of the end nodes (default = "red")}

\item{theme}{theme to apply to the final plot (default = empty)
refer to https://ggplot2.tidyverse.org/reference/theme.html for more on ggplot2 themes}
}
\description{
This function creates a plot of the decision tree with customizable appearance.
}
\examples{
tree <- plot_tree(read.csv("branches.csv"))

}
